requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
aiosqlite>=0.19.0
redis>=5.0.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Document Processing
docx2txt>=0.8
pypandoc>=1.12
reportlab>=4.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
markdown>=3.5.0

# PPT Generation (新增)
jinja2>=3.0.0
aiofiles>=23.0.0
xmltodict>=0.13.0

# Text-to-Speech (Optional)
edge-tts>=6.1.0

# Vector Search (Memory or File Storage)
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
chromadb>=0.4.0

# LLM and AI
langgraph>=0.0.20
openai>=1.0.0
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
unstructured>=0.10.0

# OCR and Document Processing
ocrmypdf>=15.0.0
tesseract>=0.1.3
portalocker>=2.7.0

# Search Tools
arxiv>=2.1.0
wikipedia>=1.4.0
aiohttp>=3.9.0
duckduckgo-search>=5.0.0
trafilatura>=1.7.0

# Configuration and Tools
python-dotenv>=1.0.0
PyYAML>=6.0.1

# System Monitoring
psutil>=5.9.0

# Development Tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0